add[BE]: Fix joinQueue method
